corrections dump-text

+++ PROGRAMMES/script_final.sh
@@ -40,9 +40,9 @@ REP_DT=../DUMP-TEXT
 REP_CT=../CONTEXTES
 
 #Vidange des répertoires
-rm $REP_PA/*.html >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_PA" || log_failure "Error:"`cat $REP_LOG/command.log`
-rm $REP_DT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_DT" || log_failure "Error:"`cat $REP_LOG/command.log`
-rm $REP_CT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_CT" || log_failure "Error:"`cat $REP_LOG/command.log`
+[ "$(ls -A $REP_PA)" ] && rm $REP_PA/*.html >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_PA" || log_failure "Error:"`cat $REP_LOG/command.log`
+[ "$(ls -A $REP_DT)" ] && rm $REP_DT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_DT" || log_failure "Error:"`cat $REP_LOG/command.log`
+[ "$(ls -A $REP_CT)" ] && rm $REP_CT/*.txt >$REP_LOG/command.log 2>&1 && log_success "Purge: $REP_CT" || log_failure "Error:"`cat $REP_LOG/command.log`
 #exit 0
 #-------------------------------------------------------
 # en-tête du fichier html
@@ -102,19 +102,18 @@ do
             #détection de l'ENCODAGE    
             detection_encodage;
 
+            #
+            dump_text;
+
             # si encodage = UTF8 alors on fait certains traitements
 				# sinon il faudra aussi les faire et probablement d'autres
             gestion_encodage;
-
-            #
-            dump_text;
 
             #
             comptage_motifs;
 
             #
             extraire_contextes;
-
 
             #
             calcul_index;
